Create README and add badges

+++ api/v1_0/views.py
@@ -4,7 +4,8 @@ from flask import g, request
 
 from ..auth import auth
 from api_init import api
-from ..decorators import json, paginate
+from ..decorators import json, paginate, authorized
+from .. import errors
 from ..models import db, BucketList, BucketListItem
 
 
@@ -18,12 +19,9 @@ def create_bucketlist():
     bucketlist.date_created = now
     bucketlist.date_modified = now
     g.user.bucketlist.append(bucketlist)
-    try:
-        db.session.add(bucketlist)
-        db.session.add(g.user)
-        db.session.commit()
-    except:
-        db.session.rollback()
+    db.session.add(bucketlist)
+    db.session.add(g.user)
+    db.session.commit()
     return bucketlist, 201, {'Location': bucketlist.get_url()}
 
 
@@ -32,7 +30,6 @@ def create_bucketlist():
 @paginate()
 def get_bucketlists():
     """Returns all bucketlists belonging to calling user"""
-    import pdb; pdb.set_trace()
     if request.args.get('q'):
         return BucketList.query.filter_by(creator_id=g.user.id).filter(
             BucketList.name.contains(request.args.get('q')))
@@ -43,14 +40,20 @@ def get_bucketlists():
 @api.route('/bucketlists/<int:id>', methods=['GET'])
 @auth.login_required
 @json
+@authorized
 def get_bucketlist(id):
     """Returns a single bucketlist"""
-    return BucketList.query.get_or_404(id)
+    bucketlist = BucketList.query.get_or_404(id)
+    if bucketlist.creator_id == g.user.id:
+        return bucketlist
+    else:
+        return errors.unauthorized()
 
 
 @api.route('/bucketlists/<int:id>', methods=['PUT'])
 @auth.login_required
 @json
+@authorized
 def edit_bucketlist(id):
     """
     Edit the name of a bucketlist
@@ -71,6 +74,7 @@ def edit_bucketlist(id):
 @api.route('/bucketlists/<int:id>', methods=['DELETE'])
 @auth.login_required
 @json
+@authorized
 def delete_bucketlist(id):
     """
     Delete a bucketlist
@@ -89,6 +93,7 @@ def delete_bucketlist(id):
 @api.route('/bucketlists/<int:id>/items/', methods=['POST'])
 @auth.login_required
 @json
+@authorized
 def create_bucketlist_item(id):
     """
     Add an item to a bucketlist
@@ -113,6 +118,7 @@ def create_bucketlist_item(id):
 @api.route('/bucketlists/<int:id>/items/<int:item_id>', methods=['GET'])
 @auth.login_required
 @json
+@authorized
 def get_bucketlist_item(id, item_id):
     """
     Returns a bucketlist item
@@ -123,12 +129,18 @@ def get_bucketlist_item(id, item_id):
     Returns:
         BucketList item query
     """
-    return BucketListItem.query.get_or_404(item_id)
+    bucketlist_item = BucketListItem.query.get_or_404(item_id)
+    if bucketlist_item.bucketlist_id == id:
+        return bucketlist_item
+    else:
+        return errors.forbidden(
+            "You do not have permission toaccess this resource")
 
 
 @api.route('/bucketlists/<int:id>/items/<int:item_id>', methods=['PUT'])
 @auth.login_required
 @json
+@authorized
 def edit_bucketlist_item(id, item_id):
     """
     Edit bucketlist item
@@ -140,6 +152,9 @@ def edit_bucketlist_item(id, item_id):
         a dictionary of the bucketlist item
     """
     bucketlist_item = BucketListItem.query.get_or_404(item_id)
+    if bucketlist_item.bucketlist_id != id:
+        return errors.forbidden(
+            "You do not have permission to access this resource")
     bucketlist_item.from_json(request.json)
     bucketlist_item.date_modified = datetime.datetime.now()
     db.session.add(bucketlist_item)
@@ -150,6 +165,7 @@ def edit_bucketlist_item(id, item_id):
 @api.route('/bucketlists/<int:id>/items/<int:item_id>', methods=['DELETE'])
 @auth.login_required
 @json
+@authorized
 def delete_bucketlist_item(id, item_id):
     """
     Delete a bucketlist item
@@ -160,6 +176,9 @@ def delete_bucketlist_item(id, item_id):
         a dictionary of the result status
     """
     bucketlist_item = BucketListItem.query.get_or_404(item_id)
+    if bucketlist_item.bucketlist_id != id:
+        return errors.forbidden(
+            "You do not have permission to access this resource")
     db.session.delete(bucketlist_item)
     db.session.commit()
     return {'result': "Successful"}
